keep extension in the same package
