--- packages/menu-filebrowser-extension/package.json
@@ -1,40 +1,0 @@
-{
-  "name": "@jupyterlab/filebrowser-extension",
-  "version": "0.10.0",
-  "description": "JupyterLab - Filebrowser Menu Extension",
-  "main": "lib/index.js",
-  "types": "lib/index.d.ts",
-  "files": [
-    "lib/*.d.ts",
-    "lib/*.js"
-  ],
-  "directories": {
-    "lib": "lib/"
-  },
-  "dependencies": {
-    "@jupyterlab/application": "^0.10.0",
-    "@jupyterlab/filebrowser-extension": "^0.10.0"
-  },
-  "devDependencies": {
-    "rimraf": "^2.5.2",
-    "typescript": "~2.4.1"
-  },
-  "scripts": {
-    "build": "tsc",
-    "clean": "rimraf lib",
-    "watch": "tsc -w"
-  },
-  "jupyterlab": {
-    "extension": true
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/jupyterlab/jupyterlab.git"
-  },
-  "author": "Project Jupyter",
-  "license": "BSD-3-Clause",
-  "bugs": {
-    "url": "https://github.com/jupyterlab/jupyterlab/issues"
-  },
-  "homepage": "https://github.com/jupyterlab/jupyterlab"
-}
